archive/v5/backend/requirements.txt
fastapi==0.128.4
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.46
alembic==1.18.3
pydantic-settings==2.12.0
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.22
psycopg[binary]==3.2.9
pytest==9.0.2
qrcode[pil]==8.2
reportlab==4.4.4
openpyxl==3.1.5
oss2>=2.18.5
